question

+++ src/main/java/com/example/reportsystem/service/RoomService.java
@@ -6,6 +6,6 @@ import org.springframework.http.ResponseEntity;
 public interface RoomService {
     ResponseEntity<?> findAll();
     ResponseEntity<?> save(RoomRequest request);
-    ResponseEntity<?> deleteById(long id);
-    ResponseEntity<?> updateById(RoomRequest request,long id);
+    ResponseEntity<?> deleteById(Integer id);
+    ResponseEntity<?> updateById(RoomRequest request,Integer id);
 }
